Initial Commit. Does not include README.md or seed.sql yet.
Once those are added, it should be easier to setup a new project. Once base is configured, we can push a final
commit and call it base. for now we are on master..

Once the final commit is done we should consider forking or creating a new project without just the CMS + base pages.